+++ ambit2-all/ambit2-fp/src/main/java/ambit2/descriptors/simmatrix/SimilarityMatrixBitset.java
@@ -12,7 +12,7 @@ public class SimilarityMatrixBitset extends SimilarityMatrix<BitSet> {
 	}
 
 	@Override
-	protected boolean accept(float t, double threshold) {
+	protected boolean accept(float t, float threshold) {
 		return (t >= threshold);
 	}
